fix: populate host header during request generation
Go's client adds in host header (if not present) using request's URL.
hit doesn't see this header since it is only populated during request
executiong i.e. when the request is written on the wire.

With this patch, the host header is populated when model.Request is
generated from a parsed request. This ensures that the console output
and the save request have the host header.

Additionally, the request headers are now populated using appropriate
http.Header.Add method rather than cloning a type casted
map[string][]string. This ensures that the headers are correctly
canonicalized in the underlying map.