feat: body is now enclosed with '~' lines (#25)
The body of the request now must start with a `~` as the first line with
an optional body encoding and end with a line containing `~`. These
lines are not sent on the wire.

The change was done to make hit files a bit more expressive
This is a breaking change for all hit users.